backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.3
celery==5.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.8
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-jet-reboot==1.3.10
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-api-python-client==1.4.1
httplib2==0.31.0
idna==3.11
kombu==5.5.4
oauth2client==4.1.3
packaging==25.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0

# Celery para tareas asíncronas
celery==5.4.0
redis==5.2.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Social Auth para Google OAuth
social-auth-app-django==5.4.2
social-auth-core==4.5.4
